refactor(event-detail-v2): re-import recommendation from _shared (step 2/2)
- EventDetail/adapters.ts: 추천 관련 export 4개 + 내부 helper 일체 제거
- EventDetail/hooks.ts: 추천 심볼 import 출처를 _shared/recommendation 으로 분리
- RecommendedSection.tsx / RecommendedCard.tsx: import 경로 갱신

승격(step 1) 커밋 8eb9ed7 와 함께 한 단위 — 빌드 무결.
Cart.plan.md § 10.1 step 4.
