+++ src/pages-v2/EventDetail/adapters.ts
@@ -3,6 +3,11 @@ import { toStatus, toDateTimeLabels, isFree, isLowStock } from '@/pages-v2/_shar
 import type { EventStatus } from '@/types-v2/event';
 import type { EventDetailVM } from './types';
 
+/**
+ * 추천 카드 어댑터(`RecommendedCardVM`, `toRecommendedCards`, ...)는
+ * `@/pages-v2/_shared/recommendation`으로 승격됨 (Cart 와 공유 — Cart.plan.md § 9.2-17).
+ */
+
 const toTechStackNames = (items: TechStackItem[]): string[] =>
   items.map((t) => t.name);
 
@@ -35,57 +40,3 @@ export const toEventDetailVM = (api: EventDetailResponse): EventDetailVM => {
   };
 };
 
-/* Recommended events (§3 표 2-B). Narrow types live here — `src/api/types.ts`
- * has no formal RecommendationResponse and SPEC §0 keeps the api/types
- * file untouched. */
-
-export interface RawRecommendedEvent {
-  eventId: string;
-  title: string;
-  price?: number;
-  eventDateTime?: string;
-  category?: string;
-}
-
-export interface RecommendationResponse {
-  events?: RawRecommendedEvent[];
-}
-
-export interface RecommendedCardVM {
-  eventId: string;
-  title: string;
-  category: string;
-  price: number;
-  isFree: boolean;
-  eventDateTime?: string;
-  dateLabel: string;
-}
-
-const RECOMMEND_LIMIT = 5;
-const FALLBACK_CATEGORY = '추천';
-const FALLBACK_DATE_LABEL = '일정 확인';
-
-const toRecommendedCardVM = (raw: RawRecommendedEvent): RecommendedCardVM => {
-  const price = raw.price ?? 0;
-  const dateLabel = raw.eventDateTime
-    ? (toDateTimeLabels(raw.eventDateTime).dateLabel || FALLBACK_DATE_LABEL)
-    : FALLBACK_DATE_LABEL;
-  return {
-    eventId: raw.eventId,
-    title: raw.title,
-    category: raw.category ?? FALLBACK_CATEGORY,
-    price,
-    isFree: isFree(price),
-    eventDateTime: raw.eventDateTime,
-    dateLabel,
-  };
-};
-
-export const toRecommendedCards = (
-  api: RecommendationResponse,
-  currentEventId: string,
-): RecommendedCardVM[] =>
-  (api.events ?? [])
-    .filter((e) => e.eventId !== currentEventId)
-    .slice(0, RECOMMEND_LIMIT)
-    .map(toRecommendedCardVM);
